test(tenant-scope): split the HTTP isolation suite to fit the per-file budget
`tests/unit/test_tenant_scope_http_isolation.py` grew to 140 tests, each
building a full app. The per-file runner gives every file a 300s subprocess
budget; the file took ~200s locally and roughly twice that on the macOS
runner, so it timed out and took the whole shard -- and main -- red. Two
consecutive runs failed identically, so this was not a flake.

The cost is teardown, not the tests: `call` totals 1.1s across all 140,
while teardown totals 157s. A full `gc.collect()` runs after every test and
walks the whole live heap, which reaches ~4.5M objects and 1.23GB RSS by the
end of the file, so each successive teardown costs more than the last --
0.19s for the first test, 1.94s for the last.

Three modules sharing one conftest, split along the boundaries the file
already documented, put each one far inside the budget:

    test_tenant_selection          57 tests   39s
    test_scoped_cost               22 tests   17s
    test_scoped_reads_and_batch    61 tests   43s

Same 140 tests, same assertions. `pytestmark` stays in each module rather
than moving to the conftest: a conftest's `pytestmark` does not reach the
modules beside it, and losing `auth_enabled` here silently disables the
authentication these tests exist to exercise.
